MT#56647 kernel-module: Add MODULE_ALIAS for ipt and ip6t iptables support
These are the entries the kernel will try to look for when userspace
tries to use the RTPENGINE support, but the module is called
xt_RTPENGINE so the kernel does not find it and thus cannot autoload
it.

The problem is that we have startup code that sets up iptables rules,
and will fail if the module has not yet been loaded. By adding these
aliases we are no longer concerned about the ordering, and can let the
kernel autoload the modules when they are needed.

Change-Id: I43a54af4e32adf755538c57df711da048c720030