MT#56128 Temporarily remove the test #1142
The test number 1142:
"SDP attr manipulations - add a= line for media audio, two times"
fails on the buster builds, because of the hash tables implementation
used in the `command_values` of the `struct sdp_command`.

It doesn't necessarily keep the same order, as was used to add values.

Temporarily removing it, until the `struct sdp_command` is re-designed.

Change-Id: I077771c00c5dc8907e42d7757540e3ffa2063af6