MT#56128 sdp manipulations: improve structures design
In order to improve the efficieny of the code (make lookups
faster) use GQueues for CMD_ADD manipulations and
GHashTable for CMD_REM manipulations.

This gives the following benefits:
- faster lookup (check), if any SDP command is to be applied
  (no need to iterate through all values of commands, as it used to be)
- keep a sequence for CMD_ADD values given by option flag(s)

In order to keep the code more lightweight and clear,
add a separate struct ptr to queues and hash tables into
the call_interfaces header.

Change-Id: I7e45aca4062750c7b8959473edb410ed76cc04e7