MT#56128 Re-add test for SDP attr manipulations
The test number 1142:
"SDP attr manipulations - add a= line for media audio, two times"
used to fail on the buster builds, because of the hash tables
implementation we had in the `command_values` of the `struct sdp_command`.
It didn't necessarily keep the same order, as was used to add values.

Now as the SDP manipulations structures is re-designed, we can re-add it.

Change-Id: Icc7d030894d45c9d62e91669915eafb348d05d97