MT#56447 transfer ownership of janus_handle
As multiple handles from multiple different sessions can participate in
the same room, we need a way to look up handles from different sessions.
We therefore transfer ownership of all janus_handle objects to global
level. As these are currently not reference-counted, this necessitates
holding the global lock for the entirety of the signalling process.

Change-Id: Ibe27cb28407d1c8d7290faaaf5b6ae7c7f71679a