MT#56447 refactor janus_send_json_msg
Distinguish between two cases: Sending a message to one particular WS or
HTTP connection directly as a response to a request, and sending an
asynchronous message or event to a session. Introduce
janus_send_json_sync_response() to do the format, and update all event
notifications to use janus_send_json_async().

Change-Id: I3b5a05e8c3c61d046bcaec29006698625a2570fa