MT#57093 more granular kernel I/O structs
The unified `rtpengine_message` struct has become quite large, and most
fields are not used for most of the messages sent to the kernel module.

Use command-specific structures that only contain the required
information. Adapt the I/O code and the size checks to use the correct
sizes.

Do special handling for GET_*_STATS as these commands do both input and
output at the same time. Copy in only the input portion of the struct.
(Copy-out still copies the entire struct at this point.)

Change-Id: Ia8aec6135fd7de42ae4e62e7f2dc23804e1f0914