MT#57118 Don't destroy a whole call, if only a branch is to be destroyed
Upon calling the `monologue_delete_iter()`,  we have to take into account
an amount of forked legs we have now entirely in the call.

If there's more than 1 forked leg, don't mark the whole call for a full destruction,
since other forks might be successful (and they are not in a scope of processing now).
Only destroy this branch and associated to it monologue(s).

Change-Id: I6fdb4ba33ae7582578d84d57d91f9536d932ca47