MT#57118 Rework the call's deconstruction logic
This commit: 6f0ad0db0fc0fce370f6bcf388e9a9e69c56d81d
introduced an improved behavior of the call deconstruction
when dealing with multiple forked legs. But it also
introduced a regression related to TPCC related calls.

This is because it takes into consideration existing
`->viabranches`, which is not efficient in certain cases.

Instead, we just have to ensure, whether there are
other monologues of the call object, which are not directly
"visible" for this branch being cancelled (so they are not directly
associated to that), but they might have own associations
present, which means the call mustn't be globally destoroyed.

Change-Id: I630f2d88ef3b557af3a95816fc2703daccaff374
(cherry picked from commit 0994ffbe7579d1751c4b8482369651757f7a2705)