MT#57268 Move release closed sockets to a separate thread
To do the work more efficiently and not be dependent on
the `call_timer` runs by poller, we should move
the releasing of sockets to a separate thread, to make it
faster and not be dependent on what happens in the `call_timer`
at all. Since it has nothing to do with the call timers.

Since now we have two queues:
- thread scope (local): ports_to_release
- global one: ports_to_release_glob

`sockets_releaser()` uses the ports_to_release_glob,
meanwhile appending in the `call_timer()` happens using the
ports_to_release.

Change-Id: Iadd966ac895b2dd64f81269d4fdf5d83747fe0b7