TT#146201 obsolete TRANSCODE flag entirely
Transcoding is flow-specific, so it doesn't make sense to have a flag
for it in a call_media section. Instead we use the transcoding flag set
on the call_subscription objects (on subscribers), and set/unset a flag
on the monologue struct, depending on whether any media flows (going to
subscribers) have transcoding enabled.

Change-Id: Id671d56e56a22eaa8e56f6d449770b0c7b086cea