MT#57335 Move `stats_rate_min_max()` to a separate thread
To do the work more efficiently,
and not be dependent on the call_timer runs by poller,
we should move the `stats_rate_min_max()` to a separate thread,
to make it faster and not be dependent on what happens in the `call_timer` at all.
Since it has nothing to do with the call timers.

Change-Id: I9a39e1b63cb8741377f5af5b2d52d4f8b428a0ad