MT#57337 Move kernel stats update to a separate thread
To do the work more efficiently,
and not be dependent on the call_timer runs by poller,
we should move the ports iterations (stats update from the kernel)
functionality to a separate thread, to make it faster and
not be dependent on what happens in the `call_timer` at all.
Since it has nothing to do with the call timers.

As an additional benefit: we unload the `call_timer` runner.

Change-Id: I511529ce504ef3d29f4e9d6d731ffd470d78d27a