MT#57378 move `stats_counters_calc_rate()` to a separate thread
Similarly as for the `stats_rate_min_max()`,
move the `stats_counters_calc_rate()` functionality
to the same separate thread, so that we do the work
more efficiently and not be dependent on the call_timer
runs by poller.

Furthermore it makes more sense to keep in the `statistics.c`,
since it obviously has to do with statistics.

Additionally:

Update the test-stats due to these changes in the `call_timer()`

We have to call the `stats_counters_calc_rate()` now explicitely from
the test-stats.c, because the `call_timer()` is not anymore
responsible for providing stats counters rate calculations.

Change-Id: I1682eb76e3057f0f431c27b9633717d965313a1a