MT#57379 move `ice_slow_timer()` to a separate thread
Move the `ice_slow_timer()` functionality to a separate thread,
so that we do the work more efficiently, and not be dependent
on the call_timer runs by poller.

Furthermore it makes more sense to keep in the `ice.c`,
since it obviously has to do with ice timing.

Additionally:

Update the test-stats due to these changes in the `call_timer()`

We have to call the `ice_slow_timer()` now explicitely from
the test-stats.c, because the `call_timer()` is not anymore
responsible for providing stats counters rate calculations.

Change-Id: I03377dd59ea71c27497e1f4d30164075f05165cd