MT#57380 Deprecate inc/dec functionality of `call_timer` run duration
Deprecate increasing/decreasing functionality of
the call timer run duration, because it's not needed anymore,
most of the unnecessary functionality has been moved lately
to separate dedicated threads and the `call_timer`
is simply unloaded for now.

Additionally: remove logging with timestamps, because
we do not need to monitor it anymore.

Change-Id: I8d1fd260472528fcee6a12f48b9a92f195893787