MT#57977 use symmetric UDP
UDP packets sent in response to a UDP request should have the same
source address as the request's destination address.

This can be achieved with sockets bound to a specific address, but in
the case of ANY-bound sockets, we can use the PKTINFO mechanism to do
this.

Extend control_ng_process() to accept an extra socket address
corresponding to the local address to use. Extend the signature of the
callback function (to do the actual sending) accordingly.

Extend socket_sendiov() to be able to set the PKTINFO cmsg when sending
a packet.

Add socket_sendto_from() as a convenience wrapper.

Extend control_udp_incoming() to pass the address from
udp_buf->local_addr back to socket_sendiov().

Change-Id: Idd019fdcfd796098e7807427e6686d4b05de35d1
(cherry picked from commit 8684a1933a3aad0731aec86ca242ea2e363cbe01)