MT#55283 fix config parsing strategy
Looks like glib at some point started resetting an existing string array
that a config option is pointing to if that config option isn't present
at all. Work around this by explicitly resetting the pointed-to variable
to NULL, and then fixing up the contents on the second pass: either use
the setting from the config file or the one from the command line, and
free the other one.

Change-Id: Ida89d71e54fc30b4cce3277107e2185737f54a51