MT#57719 redis: move media to monologue linking to  `json_medias()`
Move linking of the media to its monologue, from the `json_link_medias()`
to the `json_medias()`. This is required, becase the `json_link_medias()`
already needs to operate on medias list, which is fully linked
to their monologues. Otherwise it can happen, that during the
linking of media to media in `json_link_medias()` (subscribing them)
the monologue for the other media is still not linked.

Change-Id: Ifc4dd8dff27f65e664b4fd21332a1d296d07759b