MT#55283 extend codec_tracker_update
Allow codec_tracker_update to reference an existing codec_store. Then
when a supplemental codec type needs to be generated, make it look up
the type in the existing codec_store and re-use the existing payload
type if present instead of creating a new one. This allows payload type
numbers to remain unchanged during a re-invite.

Change-Id: I9e5edd897515a5e3eb5033aa6bbf21c8667d6133