MT#55283 fix media array building
The index j we get is not the position in the output array/list, but
rather the index into the source array. Simply append each new element
to the array in order.

Also make sure we don't skip over empty elements in the array when
building the JSON list, so that the order is preserved.

Change-Id: Id6577410e114f0ddbea745977118f1bab2e38fa9
(cherry picked from commit 64724284c19b2b825252882b174d249e5eb60f95)