MT#57550 Unconfirm monologue medias using subs
In the `dialogue_unconfirm()` (previously `dialogue_unkernelize()`)
use media subscriptions to find subscribers/subscriptions
and correlated to them medias to unconfirm them.

This change is bound to the switch from monologue
subscriptions to media subscriptions.

Additionally: Refactor unkernelize/unconfirm in `call.c`

To provide a sense of the functions clearer,
the namings of the following function singatures
have been changed:
- `dialogue_unkernelize()` -> `dialogue_unconfirm()`
- `__monologue_unkernelize()` -> `__monologue_unconfirm()`

Change-Id: If8cc2515eb26dcc4ffed02a420514f98d43cc5c6