MT#57550 Don't rely on empty from-tag in `call_delete_branch()`
Due to introduction of media subscriptions model.

Rely on subscription monologue's empty tag,
while trying to detect, whether or not to destroy given monologue
but with a clause that:
- this ml is subscribed to medias of other signle ml

It's important to keep this check because
if the `delete` is done with from-tag and to-tag,
just right away after an `offer` without the to-tag
and without use of via-branch, then the call's data
can potentially remain dangling without being deleted completely.

So looking up the offer side of the call through the from-tag
and then checking, if the call has not been answered
(answer side has an empty to-tag), gives a clue whether
to delete an entire call.

Additionally, introduce a helper function:
- `ml_medias_subscribed_to_single_ml()`
which checks, whether this monologue medias are subscribed
to a single other monologue medias.

Change-Id: I2474d24ac66ce2cb12bd282f2c0df809fce7880c