MT#57550 Create a new from-monologue, if still not matched
Rework the logic of fromtag monologue detection in the
`call_get_dialogue()`.

Fristly, try to match using given from tag (if there is one).
Secondly, try to match using given viabranch.
Thirdly, try to use top-most subscription of the top-most
     media related to the totag monologue (newer behavior).
Fourthly, if still not found just create a brand-new one.

This change relates to the fact, that we stop using call monologue
subscriptions model and switch to the media based one.

Change-Id: I106bdd79f5e4754f76bd048f51c8fc211ee57410