MT#59069 refactor handling of metafile prefix
Move the code generating the recording metafile prefix out of the
"externally" visible code and logically deeper into recording.c.

Move the field from the recording struct into the call struct so that it
can be directly restored from Redis, and eliminate and mostly unused
function argument in the process.

Functional no-op.

Change-Id: I31eb3cd5864dd8138fabad0bcbd08ef18571f1a9