MT#59069 merge handling of output-dest w metadata
Handle the `output-destination` flag in the same way as the `metadata`
flag, eliminating the need to do it explicitly everywhere it's relevant.

Add a few alternative spellings of `output-destination`.

Rename the member from flags to `recording_file` to make it more clear
what it is.

Add save/restore capabilities of that field to the Redis code.

Unify printing of monologue LABEL into the same function that handles
the metadata.

Update documentation to better explain this option, and use the new
clearer name.

Change-Id: I4496341013b0ccab5b1dec026cf3a1a0ea879018