MT#58535 rtpp_flags: introduce raw flags parsing
Add support of rtpp_flags parsing for the daemon.

From now on, it's possible to parse option flags
on the daemon side instead of the kamailio module.

It's identical to what the module does, but the
difference is:
- module sends general call identification such as:
  call-id, From/To tags, viabranch using bencode
- meanwhile all generic/non-generic option flags
  are added to the `rtpp_flags` bencode member as str
- parsing of that is done as usually using the
  `call_ng_main_flags()` /  `call_ng_flags_flags()`
  and additionally using new parser `parse_rtpp_flags()`

New file implementation and header introduced:
- control_ng_flags_parser.c
- control_ng_flags_parser.h

Otherwise no functional changes, and the parsing itself
remains working following the same algorithm.

Change-Id: I59e47fa1947e2aeaa0bbf3930a0f21d9a6d669ad