MT#55283 extend SDP attribute matching
For tagged attributes using a `a=name:tag value` format, extend pattern
matching so that just `name` matches as well as `name:tag`. Only support
this for removal (not substitution) as these attributes generally can
appear multiple times, and doing a static substitution on all of them is
kinda pointless.

Change-Id: I07c7f6e9075e862902bffe679d5de8618162eb8f