MT#55283 Don't fail deleting the iptables chain if it is still in use
When base chain is "none", the admin is required to manage jumping into the
rtpengine chain. The chain can't be deleted if it is still referenced by
another rule, which is common in this configuration.

closes #1787

Change-Id: I8a72e1041a364db60870b5acececc234c8452bab
(cherry picked from commit 2e7e720c94ad086ab33e015ea9ea4b13f17a8270)