MT#58535 Be able to detect message type (request/reply)
For the sake of proper processing for From/To tags,
we need to know which message type originated interaction
with the rtpengine (from behalf of kamailo module).

This can be of two types:
- request
- reply

This can also be re-used later, if there is a need
to detect whether that was a request or reply.

P.S.: opmode and message type shouldn't be confused,
it's not the same, since SDP offer can come in the request
as well as in the reply message (late offer/answer).
So the opmode reflects what exactly is being done in terms of
rtpengine operations, meanwhile the message type shows, which
message type originated this processing.

Change-Id: I8336175d3cfcf521418ac18b71cb4a6d5730705d