MT#55283 fix missing EVS size collision avoidance
With hf-only=0 set and cmr=1 also set, we send a blank CMR with each
packet, which forces use of HF format packets. This in turn can lead to
size collisions (e.g. mode 1 = 18 bytes, + CMR + TOC = 20 bytes, which
is the same as mode 2) which can confuse the decoder. Add the required
size collision logic.

Change-Id: Icd540cfb321b4eaf77a0ed7d08be06d8681b6214