MT#55283 move interface stats to shm
Use the kernel/user shared memory to keep track of interface stats. This
eliminates the need to update the userspace stats counters with the
values from the kernel, and vice versa when updating the kernel streams.

Add a function to reverse map userspace mapped memory address to kernel
space addresses.

Change-Id: Iaa5f9488061a12103e57faf27b3979521778cea8