MT#55283 move last_packet timestamp to shm
Keep two separate timestamps, one updated by userspace code only and the
other updated by kernel only. This way we can tell where the packet
processing happens. For code that wants to report only the last
timestamp regardless of which one of the two it is, we add a convenience
function that just returns the newer one.

Change-Id: Ib3af7aa55006d8b32e2bc3db4f8bfa5514c57e40