MT#55283 media_socket.c: check for NULL payload_types passed to kernelize_one
under normal circumstances, the rtp sink enters this function first
and updates the reti->local.family var to AF_INET so that when the
function is called for an rtcp sink, it exits the function early.
However, if media is being blocked the rtcp sink is the first to
enter the function and does so with a NULL payload type. this NULL
check is therefore required to prevent a sefgault on the call to `assert`.

closes #1876

Change-Id: Id95cbbc9ae388f50dd51417c351f4a9ba09eb5f0
(cherry picked from commit 8e04c44b867bb3b6967c2df471e5204ec37d4784)
(cherry picked from commit 90a5537126538a6f5e4a0a99e5bdc41a01ee035f)