MT#60304 debian: Switch from markdown to discount
The markdown package seems dead upstream, and the packaging situation
in Debian is not very good either. It currently has an RC bug which
means rtpengine itself is marked for autoremoval from testing. While
we could fix the RC bug for the markdown package, it's probably better
to switch implementation instead to something that is more lively. That
would be the discount implementation for example.

Change-Id: I25334f1539090d5e2d8008f460a7459713241d7c