MT#60476 sdp_replace: generalize usage of `->session_last_sdp_orig`
Use the `->session_last_sdp_orig` for the version as well as
username type of replacements.

Remove superfluous code.

It's possible to handle it like that, because the last user
of the sdp_replace func, which is the offer/answer model, sets
the `->session_sdp_orig` actually in `__call_monologue_init_from_flags`.

Additionally: add tests for the user type of the origin replacement.
Change-Id: Idd60e628c676840c2e844077303035c0edec67e2