MT#60476 sdp_replace: correct behavior of `replace-origin-full`
The option flag `replace-origin-full` should replace whole
origin (o=) line of the SDP body, so that all origin fields
in the `o=` line always remain the same in all SDPs going to a
particular RTP endpoint.

So, not just a straight-forward replacing with own values.

A behavior in relation to the address field is the same
as by the `origin` option flag.

Correct according tests and add one more.

Change-Id: I4bcb740db5cd68a75fb754f5f6b5d21e19b05e1b