MT#60476 add handling of `a=setup:` for message type
For the `sdp_create` based approach, introduce
`a=setup:` handling in cases with the message media
type processing.

Currently this has to be divided for sdp_replace
and sdp_create users. This is because sdp_replace one
uses chopper based approach, which in combination with
changes in this commit dupplicates the setup attribute.

In order to separate the logic both approaches,
introduce a temporary bool, until the time when
sdp_replace gets fully deprecated.

A move of the check out of the `is_active` condition
is required, because the message media type isn't active
and hence has to be checked out of the active context.

Change-Id: I32b8f3f3928ba22e030ab5a6414b93fcd7c47201