MT#60476 sdp_create: fix crash on zero stream
It's been noticed that trying to print a zero
conneciton address, on the media stream, which
actually offered non-zeroed IP source, but a
zero port, leads to a segmentation fault,
because it relies in this case on presented
`sockaddr_t` information, which is actually absent.

This happens only with sdp_create on offer/answer.
Test fixed is "a=mid on zero streams".

Change-Id: Ic2f7f671c9ded9b361d6467ca013698fdff28538