MT#61140 call_delete_ng: add normal parsing
Add `call_ng_process_flags()` based parsing
as for other opmodes, like offer and answer.

This keeps the backwards compatibility with
the older "flags" parsing approach on the
module side, as well as adds the possibility
to parse rtpp-flags on the daemon side.

As an advantage, there is no need to use
specific local parsing for things like
to/from tags, call-id, delete-delay etc.

Additionally:
- this commit introduces flags-flags parsing
  for the "fatal" flag.
  However, as before is only taken into account
  by the `call_delete_ng()` processing,
  so no functional change.
- this commit introduces main-flags parsing
  for the "delete-delay" flag, which is also
  only taken into account by the `call_delete_ng()`
  processing, so no functional change.
- this commit adds To-tag options flag prasing
  into the `call_ng_flags_flags()` function,
  and is used by `call_delete_ng()` specificially,
  for cases when more specific identification of
  monologues to be deleted is used.

Change-Id: Ia992e5375a2f86318d9ad193a7857dd589038eed