MT#55283 switch to BIO_set_callback_arg
Use a BIO WRITE callback instead of BIO_read'ing from the BIO after each
operation. This is a more direct way to intercept data that needs to be
sent out.

Implement MTU-related BIO callbacks.

Deduct the assumed IP MTU overhead from the configured MTU during
startup.

Unlike the previous code, this does not necessarily send DTLS from the
same socket that received a message, nor to the same address that sent
one, and instead always uses the selected_sfd and ->endpoint. This may
or may not be a regression.

Closes #1806

Change-Id: I4d4456df3f378d00782cbfa64afdb2a038217e6c