MT#60476 Deprecate ATTR_LABEL parsing
Because not specificially used in rtpengine,
just deprecate it. Not to be mixed with
the `label=xxx` usage via NG protocol.

This will be also required when switching to
the sdp create approach, because needs to be
delivered via generics to print out SDP.

No functional change. Tests had to be fixed accordingly,
because of the appeared offset in the list of attributes.

Additionally, tests containing coming SDP with a label,
but which are within inactive media, don't pass it further.
One single example of it is an answer case of "a=mid mixup" test.
This is because of `print_sdp_media_section()` specifics,
and only printing generics for active media sessions.

Change-Id: Iabb44adf17adb5d8e0c5bcad20f527b591b2222c