MT#56465 make sdp manipulations helpers array agnostic
Convert existing sdp manipulation helpers
to work with arrays directly, independently from
which array is actually given.

Moreover, don't use the `G_N_ELEMENTS()`,
since we already have the `__MT_MAX`
as the size of the array, so we can directly use
that instead of relying on the G_N_ELEMENTS macro
(for array size checks).

Change-Id: I38e71dab30a467b933c213c8157fa06651ebaf04