MT#61371 support changing of to-tag
If we receive an answer from a to-tag that hasn't previously seen and no
corresponding monologue exists (created from a previous offer),
previously we would treat this is a separate and new call branch, create
a brand new monologue and dissociate the previous one. This may lead to
unexpected results as this new monologue has been created without the
same initialisation as was done for the original one, and so may be left
with incorrect or incomplete data (e.g.  SRTP keys, codec information,
interface bindings, etc).

Improve this by treating an unexpected and unseen to-tag as an alias to
the already existing to-tag. Going forward both tags can be used
interchangeably to refer to the same monologue.

Add a flag to suppress this new behaviour, in case some situation is
made worse by it.

Change-Id: Ie8f838eebd50d29d3549031998a2eb1f610b04bb