MT#61368 use media subscription lookup for offers
To support scenarios with mismatched media sections between two sides of
a call that aren't based on explicit SDP manipulations, attempt to
lookup the peer media based on an existing subscription also for offers.

Add extra checks (media type, media ID, monologue pointer) to make sure
that the obtained media object is actually the correct one.

Use a fallback on the existing index-based lookup in case no matching
media can be found.

As a side effect, the numbering and ordering of media objects created
during offer/answer is now reversed as the function is called in a
different order. Adapt tests as needed.

Change-Id: Ia5351f7d4fed03098c6be6bea3d4c74f52b8a76b