MT#61630 moh: add duration limit
Add duration limit for the moh functionality.
This is required due to a possibility to get
monologues and hence packet streams stuck
forever and sending audio to recipients
which already don't exist.

A configuration option that controls that:
`moh-max-duration` - gets a value in milliseconds.

By default is set to 1800000 (half an hour).

Change-Id: Id50a0a10ce5b52b3876a3122fb16a71accec90ff