MT#61822 improve CLI keyword matching
Doing a simple prefix string match is not enough as there may be
keywords that are equal to a prefix of another keyword. Make sure the
keyword is followed by a space or newline or end of line. This also
removes the burden of having to skip over leading spaces in the CLI
handlers.

Change-Id: I118a08e1b33d9bf592d3d80101e4f0634401adbd